+++ q2/main.c
@@ -14,7 +14,7 @@ int main(void)
     {
         aa = (a * a + 3.0) / 4.0; // 漸化式
         printf("n: %3d, x: %lf\n", n, aa);
-        error = fabsf(aa - a) / fabsf(a); // 相対誤差
+        error = fabs(aa - a) / fabs(a); // 相対誤差
         if (error < EPS) // 相対誤差判定
         {
             printf("収束しました\n");
